Enhance SB polling intervals (#4328)
* Change usage of retry-after header

With this change the retry-after header will be treated as an absolute value and not used for exponential backoff.
Whenever the retry-after interval sent by the service broker is larger than the calculated interval using the base interval and exponential backoff, the retry-after interval will be used. When the CC calculated interval is larger the retry-after interval will be ignored.

* Introduce config to cap service related polling interval

The maximum polling interval was hardcoded to 24h when set on the job. Though, the exponential backoff calculation did not take this into account.
Now the maximum can be configured through `broker_client_maximum_async_poll_interval_seconds` and will be taken into account for all scenarios. Default is still 24h and the config is optional.