Improve error handling for already deleted resources
If a user triggers two DELETE requests for the same resource within a few milliseconds two deletion jobs are created.
If both jobs are be executed in parallel it might happen that one of them fails with a `Sequel::NoExistingObject` error.
This change prevents this by catching those errors but only if the error message originates from the current model/ table.
Therefore it is ensured that similar error messages when e.g. deleting sub resources still result in job failures.

The same applies if a user unbind routes but also deletes the route itself within a very short time period.
For this case the route model now also catches `Sequel::NoExistingObject` errors related to route bindings.

Additionally, the error handling in the DeleteActionJob was improved to catch errors when the actions itself don't handle them.