Workaround changed behavior in Protobuf's Message#to_h
The behavior of 'to_h' for Protobuf objects changed [1]; unset and
default values are now being omitted. Thus the update of Protobuf
libraries [2] led to an incompatible change as the Cloud Controller
returned "null" instead of "0" for the external and internal ports as
part of the process stats endpoint.

With this change the hash is created manually by accessing all fields in
the Protobuf object(s) directly.

[1] https://github.com/protocolbuffers/protobuf/pull/15234
[2] https://github.com/cloudfoundry/cloud_controller_ng/pull/4359