Make DBs run on ramdisk and optimize rspec parallelism
Since in a dev scenario we dont care about data loss we can speed up the
tests majorly by running the DBs on a ramdisk.
Also since the DB consumes CPU during the test its not good to start
rspec executes equal to the number of CPU cores since the sume of load
excedes that by a lot. So we leave 20% spare for the DB to have less
CPU preemtive scheduling and all in all finish the rspecs faster.

An environment variable was introduced PARALLEL_TEST_PROCESSORS_MULTIPLE
that can be set to any floating point value and multiplies the CPU core
count times PARALLEL_TEST_PROCESSORS_MULTIPLE. It defaults to 0.8 and
rounds the CPU count up to the next integer so it cannot be less than 1
ever.

Apply DB performance optimizations also for github actions