Allow multiple service bindings
Allows operators to set the number of allowed service bindings via parameter `max_service_credential_bindings_per_app_service_instance`.
This will be set to 1 in capi-release to not change the current behaviour.
If set to >1 developers can create multiple bindings for the same app and service instance.
This is useful e.g. for rotating the binding credentials without pushing the app again. Restart/restage is sufficient to put the latest binding into the app.

Further details can be found in RFC-0040:
https://github.com/cloudfoundry/community/blob/main/toc/rfc/rfc-0040-service-binding-rotation.md