Move service binding uniquness validation into ccng (#4521)
* Move service binding uniqueness validation into ccng 

Allows operators to set the number of allowed service bindings via parameter `max_service_credential_bindings_per_app_service_instance`. (currently hard coded to 1)
This will be set to 1 in capi-release to not change the current behavior.
If set to >1 developers can create multiple bindings for the same app and service instance.
This is useful e.g. for rotating the binding credentials without pushing the app again. Restart/restage is sufficient to put the latest binding into the app.


To ensure users cannot run create multiple service bindings during the
bosh deployment (when old and new code is running) we need to first
introduce the new validation logic so that we can safely remove the
unique constraints in the next capi release.

Further details can be found in RFC-0040:
https://github.com/cloudfoundry/community/blob/main/toc/rfc/rfc-0040-service-binding-rotation.md

* Allow only one binding for v2 API

* fix test issues

* Remove db migration and config parameter
e.

* skip tests which require unique constraint removal

* remove config parameter

* Lock service instance to prevent concurrent binding creation