Count pending tasks towards memory and log rate limit quota
Tasks remain in state `PENDING` only for a very short period - from creation until submission to Diego (few ms).
If a user creates many tasks in parallel it might happen that the quota check will fail although the task is already submitted to Diego.
This change adds pending tasks to the remaining memory and log rate limit calculation of spaces and orgs.
In other words pending tasks behave like apps where there is only a desired started and stopped state and no sync with the actual state in Diego.

Additionally the memory and log rate limit quota calculation is now skipped if the task state changes from `PENDING` to `RUNNING` to allow updates even if the quota is exceeded.