Restore service binding ordering in sys env presenter (#4665)
The ordering of multiple service bindings for the same service with VCAP
service has changed. They are now sorted by service instance guid and
not service binding id. This was implicit behaviour.

While clients should be inspecting the binding
object itslef and not relying on the index, unfortunately not all clients
do that.

This appears to have been introduced as part of this commit to support binding rotation
https://github.com/cloudfoundry/cloud_controller_ng/commit/79b4da307661ec2a03830731414fa2c3a70d94f9#diff-403a1ff3052c3335b1d4f0db3569c895bf7ba4ca01539c562d009394c907d64bR148  and this migration https://github.com/cloudfoundry/cloud_controller_ng/commit/f01165168799634a2e90cc923db78bff1d40d542
and I believe the introduction of the index:
```
add_index %i[app_guid service_instance_guid], name: :service_bindings_app_guid_service_instance_guid_index
```

Means that the query to obtain the binding list now uses this index, and it uses the
natural order of that index, which is by service_instance_guid
rather than previously which would have been the natural order by id.

This change restores the previous behaviour by explicitly specifying the
ordering.