Convert IncompatibleBindings error to UnprocessableEntity (ApiError)
The ServiceBindingFilesBuilder can raise an IncompatibleBindings
exception in case of entries violating the specification for file names
or if the overall file size is too big. When pushing an app these errors
were shown as StagerErrors (with the correct error message); for other
operations - i.e. restart, restage - an UnknownError was returned. This
change ensures that in any case an UnprocessableEntity (ApiError) is
returned (with the concrete error message).