File-based-service-bindings: improved errors (#4693)
* Convert IncompatibleBindings error to UnprocessableEntity (ApiError)

The ServiceBindingFilesBuilder can raise an IncompatibleBindings
exception in case of entries violating the specification for file names
or if the overall file size is too big. When pushing an app these errors
were shown as StagerErrors (with the correct error message); for other
operations - i.e. restart, restage - an UnknownError was returned. This
change ensures that in any case an UnprocessableEntity (ApiError) is
returned (with the concrete error message).

* Fix validation of conflicting manifest features

Switching from service-binding-k8s to file-based-vcap-services via
manifest entries (features) raised an error. This change fixes the wrong
validation for this corner case.